requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
gunicorn==23.0.0
numpy
openai==1.59.7
PyPDF2==3.0.1
regex
requests==2.32.3
sqlalchemy-orm==1.2.10
levenshtein~=0.26.1
MarkupSafe~=2.1.5
pip~=25.1
SQLAlchemy~=2.0.31
pandas~=2.2.2
python-dotenv~=1.0.1
scikit-learn~=1.5.0
psycopg2-binary
pydantic==1.10.17
pydantic_core
scipy
elasticsearch==8.10.0
bs4
bcrypt
markdown
redis
flask_session
flask_sqlalchemy
flask-cors
gunicorn
flask-dance
stripe>=5.0.0
rapidfuzz
sendgrid==6.11.0
sentry_sdk